+++ src/libtrx/game/game_flow/reader.c
@@ -101,6 +101,26 @@ static void M_LoadCommonSettings(
             settings->fog_end.value = value;
         }
     }
+
+    {
+        JSON_ARRAY *const tmp_arr = JSON_ObjectGetArray(obj, "water_color");
+        if (tmp_arr != nullptr) {
+            const RGB_F color = {
+                JSON_ArrayGetDouble(tmp_arr, 0, JSON_INVALID_NUMBER),
+                JSON_ArrayGetDouble(tmp_arr, 1, JSON_INVALID_NUMBER),
+                JSON_ArrayGetDouble(tmp_arr, 2, JSON_INVALID_NUMBER),
+            };
+            if (color.r != JSON_INVALID_NUMBER && color.g != JSON_INVALID_NUMBER
+                && color.b != JSON_INVALID_NUMBER) {
+                settings->water_color.is_present = true;
+                settings->water_color.value = (RGB_888) {
+                    color.r * 255.0f,
+                    color.g * 255.0f,
+                    color.b * 255.0f,
+                };
+            }
+        }
+    }
 }
 
 static void M_LoadCommonRoot(JSON_OBJECT *const obj, GAME_FLOW *const gf)