+++ src/tr1/game/level.h
@@ -2,11 +2,7 @@
 
 #include "game/game_flow/types.h"
 
-#include <libtrx/colors.h>
+#include <libtrx/game/level.h>
 
 bool Level_Initialise(const GF_LEVEL *level, GF_SEQUENCE_CONTEXT seq_ctx);
 void Level_Load(const GF_LEVEL *level);
-
-RGB_888 Level_GetWaterColor(void);
-float Level_GetFogStart(void);
-float Level_GetFogEnd(void);