+++ src/tr2/game/level.h
@@ -2,12 +2,9 @@
 
 #include "game/game_flow/types.h"
 
-#include <libtrx/game/level/common.h>
+#include <libtrx/game/level.h>
 
 void Level_Init(void);
 bool Level_Initialise(const GF_LEVEL *level, GF_SEQUENCE_CONTEXT seq_ctx);
 bool Level_Load(const GF_LEVEL *level);
 void Level_Unload(void);
-
-float Level_GetFogStart(void);
-float Level_GetFogEnd(void);